Polished code for Server-Client communication
Before this, the server had a predefined response that it would always respond to the client with. Now every time it receives a request from the client, it will ask the user to type a response in the terminal which it will use to send to the client.

Also got rid of the monitorQuit() method to easily implement Server-Client communication. Otherwise, the user would have to type their response twice.